Add GitHub Actions CI workflow and add conformance testing (#4)
* Add GitHub Actions CI workflow

* Add protobuf conformance tests and update CI workflow

* Fix fmt CI job: scope packages and generate conformance code

* Fix fmt CI job: sort generated mod.rs and restore Cargo.toml

* Fix 3 runtime conformance bugs and populate failing_tests.txt

* Add unit tests for sint32 zigzag

* Expand failing_tests.txt with proto2 and missing proto3 failures

* Fix ZST remainder panic and conformance runner cache key

+++ quick-protobuf/src/sizeofs.rs
@@ -50,7 +50,7 @@ pub fn sizeof_uint64(v: u64) -> usize {
 
 /// Computes the binary size of the varint encoded sint32
 pub fn sizeof_sint32(v: i32) -> usize {
-    sizeof_varint(((v << 1) ^ (v >> 31)) as u64)
+    sizeof_varint(((v << 1) ^ (v >> 31)) as u32 as u64)
 }
 
 /// Computes the binary size of the varint encoded sint64
